+++ CHANGELOG.md
@@ -12,7 +12,8 @@ app.
   ([#978](https://github.com/birchill/10ten-ja-reader/issues/978)).
 - Made non-matching kanji headwords be hidden when matching on kana
   ([#979](https://github.com/birchill/10ten-ja-reader/issues/979)).
-- Handle search-only dictionary headwords.
+- Added handling for search-only dictionary headwords.
+- Made it possible to copy name preview entries to the clipboard.
 - Fixed display of currency conversion etc. when there are only name results
   ([#970](https://github.com/birchill/10ten-ja-reader/issues/970)).